feat: semantic search example using transformer.js
