+++ src/screens/coding-standards/coding-standards.store.utils.ts
@@ -1,34 +1,10 @@
 import { isBlank, isNullish } from "@lichens-innovation/ts-common";
-import type { EmbeddingsProgress, GuidelineNode, GuidelineSource, Rule } from "./coding-standards.types";
+import type { EmbeddingsProgress, GuidelineNode, Rule } from "./coding-standards.types";
 import type { EmbeddingsEngine } from "./utils/embeddings-engine";
 import { combineSearchResults, filterGuidelines } from "./utils/search.utils";
 
-export const DEFAULT_GUIDELINE_SOURCES: GuidelineSource[] = [
-  {
-    id: "react-patterns",
-    name: "React Patterns",
-    url: "https://raw.githubusercontent.com/amwebexpert/chrome-extensions-collection/master/packages/coding-guide-helper/public/markdowns/common-react-patterns.md",
-    enabled: true,
-  },
-  {
-    id: "typescript-patterns",
-    name: "TypeScript Patterns",
-    url: "https://raw.githubusercontent.com/amwebexpert/chrome-extensions-collection/master/packages/coding-guide-helper/public/markdowns/common-coding-patterns.md",
-    enabled: true,
-  },
-  {
-    id: "testing-patterns",
-    name: "Testing Patterns",
-    url: "https://raw.githubusercontent.com/amwebexpert/chrome-extensions-collection/master/packages/coding-guide-helper/public/markdowns/common-unit-testing.md",
-    enabled: true,
-  },
-  {
-    id: "naming-patterns",
-    name: "Naming Patterns",
-    url: "https://raw.githubusercontent.com/amwebexpert/chrome-extensions-collection/master/packages/coding-guide-helper/public/markdowns/common-naming-patterns.md",
-    enabled: true,
-  },
-];
+export const MAX_RESULTS_FOR_SEMANTIC_SEARCH = 10;
+
 export const INITIAL_EMBEDDINGS_PROGRESS: EmbeddingsProgress = {
   isCompleted: false,
   total: 0,
@@ -44,12 +20,12 @@ interface RunSearchArgs {
 
 export const runSearch = async ({ query, rootNode, embeddingsEngine }: RunSearchArgs): Promise<Rule[]> => {
   if (isBlank(query) || isNullish(rootNode)) return [];
+  const isReady = embeddingsEngine?.isReadyForSemanticSearch === true;
 
   const exactMatches = filterGuidelines({ search: query, rootNode });
-  const semanticMatches =
-    embeddingsEngine?.isReadyForSemanticSearch === true
-      ? await embeddingsEngine.findRelevantDocuments({ queryText: query, maxResults: 10 })
-      : [];
+  const semanticMatches = isReady
+    ? await embeddingsEngine.findRelevantDocuments({ queryText: query, maxResults: MAX_RESULTS_FOR_SEMANTIC_SEARCH })
+    : [];
   return combineSearchResults({ exactMatches, semanticMatches });
 };
